ecosmetics/requirements.txt
asgiref==3.7.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
Faker==25.0.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1